# generated/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/me": {
      "get": {
        "operationId": "CurrentUserController_getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current authenticated user identity, roles, and resolved app access.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "user",
                    "access"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "user": {
                      "type": "object",
                      "required": [
                        "id",
                        "email",
                        "displayName",
                        "preferredLanguage",
                        "status",
                        "roles"
                      ],
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "preferredLanguage": {
                          "type": "string",
                          "nullable": true
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "active",
                            "inactive",
                            "invited",
                            "archived"
                          ]
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "enum": [
                              "CANDIDATE",
                              "BROTHER",
                              "OFFICER",
                              "SUPER_ADMIN"
                            ]
                          }
                        }
                      }
                    },
                    "access": {
                      "type": "object",
                      "required": [
                        "mobileMode",
                        "adminLite",
                        "candidateOrganizationUnitId",
                        "memberOrganizationUnitIds",
                        "officerOrganizationUnitIds",
                        "approval"
                      ],
                      "additionalProperties": false,
                      "properties": {
                        "mobileMode": {
                          "type": "string",
                          "enum": [
                            "public",
                            "candidate",
                            "brother"
                          ]
                        },
                        "adminLite": {
                          "type": "boolean"
                        },
                        "candidateOrganizationUnitId": {
                          "type": "string",
                          "format": "uuid",
                          "nullable": true
                        },
                        "memberOrganizationUnitIds": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "uuid"
                          }
                        },
                        "officerOrganizationUnitIds": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "uuid"
                          }
                        },
                        "approval": {
                          "type": "object",
                          "nullable": true,
                          "required": [
                            "state",
                            "expiresAt",
                            "scopeOrganizationUnitId"
                          ],
                          "additionalProperties": false,
                          "properties": {
                            "state": {
                              "type": "string",
                              "enum": [
                                "pending",
                                "rejected",
                                "expired"
                              ]
                            },
                            "expiresAt": {
                              "type": "string",
                              "format": "date-time",
                              "nullable": true
                            },
                            "scopeOrganizationUnitId": {
                              "type": "string",
                              "format": "uuid",
                              "nullable": true
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Authentication is missing or the authenticated user is inactive/archived.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "details",
                        "requestId",
                        "timestamp"
                      ],
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {}
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/candidate/dashboard": {
      "get": {
        "operationId": "CandidateDashboardController_getDashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Candidate dashboard payload scoped to the active candidate profile.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "profile",
                    "nextStep",
                    "upcomingEvents",
                    "announcements"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "profile": {
                      "type": "object",
                      "required": [
                        "id",
                        "userId",
                        "displayName",
                        "email",
                        "preferredLanguage",
                        "status",
                        "assignedOrganizationUnit",
                        "responsibleOfficer"
                      ],
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "userId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "displayName": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 200
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "maxLength": 320
                        },
                        "preferredLanguage": {
                          "type": "string",
                          "nullable": true,
                          "minLength": 2,
                          "maxLength": 10
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "active"
                          ]
                        },
                        "assignedOrganizationUnit": {
                          "nullable": true,
                          "oneOf": [
                            {
                              "type": "object",
                              "required": [
                                "id",
                                "name",
                                "city",
                                "country",
                                "parish"
                              ],
                              "additionalProperties": false,
                              "properties": {
                                "id": {
                                  "type": "string",
                                  "format": "uuid"
                                },
                                "name": {
                                  "type": "string",
                                  "minLength": 1,
                                  "maxLength": 200
                                },
                                "city": {
                                  "type": "string",
                                  "minLength": 1,
                                  "maxLength": 200
                                },
                                "country": {
                                  "type": "string",
                                  "minLength": 1,
                                  "maxLength": 200
                                },
                                "parish": {
                                  "type": "string",
                                  "nullable": true,
                                  "minLength": 1,
                                  "maxLength": 200
                                }
                              }
                            }
                          ]
                        },
                        "responsibleOfficer": {
                          "nullable": true,
                          "oneOf": [
                            {
                              "type": "object",
                              "required": [
                                "id",
                                "displayName",
                                "email",
                                "phone"
                              ],
                              "additionalProperties": false,
                              "properties": {
                                "id": {
                                  "type": "string",
                                  "format": "uuid"
                                },
                                "displayName": {
                                  "type": "string",
                                  "minLength": 1,
                                  "maxLength": 200
                                },
                                "email": {
                                  "type": "string",
                                  "format": "email",
                                  "maxLength": 320
                                },
                                "phone": {
                                  "type": "string",
                                  "nullable": true,
                                  "minLength": 1,
                                  "maxLength": 40
                                }
                              }
                            }
                          ]
                        }
                      }
                    },
                    "nextStep": {
                      "type": "object",
                      "required": [
                        "id",
                        "label",
                        "body",
                        "targetRoute",
                        "priority"
                      ],
                      "additionalProperties": false,
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 80
                        },
                        "label": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 120
                        },
                        "body": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 1000
                        },
                        "targetRoute": {
                          "type": "string",
                          "enum": [
                            "CandidateContact",
                            "CandidateRoadmap",
                            "CandidateEvents"
                          ]
                        },
                        "priority": {
                          "type": "string",
                          "enum": [
                            "normal",
                            "attention"
                          ]
                        }
                      }
                    },
                    "upcomingEvents": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "title",
                          "type",
                          "startAt",
                          "endAt",
                          "locationLabel",
                          "visibility"
                        ],
                        "additionalProperties": false,
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "title": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 200
                          },
                          "type": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 80
                          },
                          "startAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "endAt": {
                            "type": "string",
                            "nullable": true,
                            "format": "date-time"
                          },
                          "locationLabel": {
                            "type": "string",
                            "nullable": true,
                            "minLength": 1,
                            "maxLength": 200
                          },
                          "visibility": {
                            "type": "string",
                            "enum": [
                              "PUBLIC",
                              "FAMILY_OPEN",
                              "CANDIDATE",
                              "ORGANIZATION_UNIT"
                            ]
                          }
                        }
                      }
                    },
                    "announcements": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "title",
                          "body",
                          "publishedAt"
                        ],
                        "additionalProperties": false,
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "title": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 200
                          },
                          "body": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 2000
                          },
                          "publishedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Authentication, approval, or an active candidate profile is required.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "details",
                        "requestId",
                        "timestamp"
                      ],
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "items": {}
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "candidate"
        ]
      }
    },
...
  },
  "info": {
    "title": "JP2 Knights API",
    "description": "V1 REST API for the JP2 Knights app ecosystem.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
...
}